restructured, runs integration-test for jetty-jersey
