+++ samples/zmon-actuator-jetty-jersey/pom.xml
@@ -8,15 +8,28 @@
 	</parent>
 	<artifactId>zmon-actuator-jetty-jersey</artifactId>
 
-	<properties>
-		<spring-boot.version>1.3.0.RC1</spring-boot.version>
-	</properties>
-
 	<dependencies>
 		<dependency>
 			<groupId>org.zalando.zmon</groupId>
 			<artifactId>zmon-actuator</artifactId>
 			<version>${project.version}</version>
+		</dependency>
+		<dependency>
+			<groupId>org.zalando.zmon</groupId>
+			<artifactId>zmon-actuator-jaxrs-jersey</artifactId>
+			<version>${project.version}</version>
+		</dependency>
+
+		<!-- MAKE SURE, THIS -->
+		<dependency>
+			<groupId>org.glassfish.jersey.core</groupId>
+			<artifactId>jersey-server</artifactId>
+			<version>2.22.1</version>
+		</dependency>
+		<dependency>
+			<groupId>org.glassfish.jersey.containers</groupId>
+			<artifactId>jersey-container-servlet-core</artifactId>
+			<version>2.22.1</version>
 		</dependency>
 
 		<dependency>
@@ -59,7 +72,12 @@
 			<groupId>org.assertj</groupId>
 			<artifactId>assertj-core</artifactId>
 			<scope>test</scope>
-			<version>3.1.0</version>
+			<version>2.0.0</version>
+		</dependency>
+		<dependency>
+			<groupId>commons-httpclient</groupId>
+			<artifactId>commons-httpclient</artifactId>
+			<version>3.1</version>
 		</dependency>
 	</dependencies>
 </project>
