+++ samples/zmon-actuator-jetty-jersey/src/main/java/org/zalando/zmon/boot/jetty/config/JerseyConfig.java
@@ -1,24 +1,26 @@
+/**
+ * Copyright (C) 2015 Zalando SE (http://tech.zalando.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package org.zalando.zmon.boot.jetty.config;
 
-import java.io.IOException;
-import java.lang.reflect.Method;
-
 import javax.annotation.PostConstruct;
-import javax.ws.rs.Path;
-import javax.ws.rs.container.ContainerRequestContext;
-import javax.ws.rs.container.ContainerRequestFilter;
 
-import org.glassfish.jersey.server.ContainerRequest;
-import org.glassfish.jersey.server.ExtendedUriInfo;
 import org.glassfish.jersey.server.ResourceConfig;
-import org.glassfish.jersey.server.internal.process.Endpoint;
-import org.glassfish.jersey.server.internal.routing.UriRoutingContext;
-import org.glassfish.jersey.server.model.ResourceMethodInvoker;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.jersey.JerseyProperties;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.core.annotation.AnnotationUtils;
-import org.zalando.zmon.boot.jetty.controller.SimpleController;
+import org.zalando.zmon.boot.jetty.resources.SimpleResource;
+import org.zalando.zmon.jaxrs.jersey.BestMatchingPatternFilter;
 
 /**
  * 
@@ -28,85 +30,11 @@ import org.zalando.zmon.boot.jetty.controller.SimpleController;
 @Configuration
 public class JerseyConfig extends ResourceConfig {
 
-	@Autowired
-	private JerseyProperties jerseyProperties;
-
-	public JerseyConfig() {
-	}
-
 	@PostConstruct
 	public void init() {
-		register(SimpleController.class);
-		register(new RequestFilter(jerseyProperties.getApplicationPath()));
-	}
-
-	static class RequestFilter implements ContainerRequestFilter {
-
-		private final String prefix;
-
-		public RequestFilter(String prefix) {
-			this.prefix = prefix;
-		}
-
-		public RequestFilter() {
-			this("");
-		}
-
-		public void filter(ContainerRequestContext requestContext) throws IOException {
-			ContainerRequest containerRequest = (ContainerRequest) requestContext;
-			ExtendedUriInfo extendedUriInfo = containerRequest.getUriInfo();
-			UriRoutingContext uriRoutingContext = (UriRoutingContext) extendedUriInfo;
-			Endpoint endpoint = uriRoutingContext.getEndpoint();
-			ResourceMethodInvoker invoker = (ResourceMethodInvoker) endpoint;
-			ResourceInfo info = getResourceInfo(invoker);
-			String matchedPath = info.getMatchedPath();
-			containerRequest.setProperty("org.springframework.web.servlet.HandlerMapping.bestMatchingPattern",
-					matchedPath);
-			System.out.println("MATCHED PATH : " + matchedPath);
-		}
-
-		protected ResourceInfo getResourceInfo(ResourceMethodInvoker invoker) {
-			return new ResourceInfo(invoker.getResourceClass(), invoker.getResourceMethod(), prefix);
-		}
-
-	}
-
-	static class ResourceInfo {
-
-		private final Method method;
-		private final Class<?> clazz;
-		private final String prefix;
-
-		public ResourceInfo(Class<?> clazz, Method method, String prefix) {
-			this.clazz = clazz;
-			this.method = method;
-			this.prefix = prefix;
-		}
-
-		public Method getMethod() {
-			return method;
-		}
-
-		public Class<?> getClazz() {
-			return clazz;
-		}
-
-		public String getMatchedPath() {
-			StringBuilder sb = new StringBuilder();
-			sb.append(prefix);
-			sb.append(getClassPathAnnotation());
-			sb.append(getMethodPathAnnotation());
-			return sb.toString();
-		}
-
-		private String getClassPathAnnotation() {
-			Path path = AnnotationUtils.findAnnotation(clazz, Path.class);
-			return path != null ? path.value() : "";
-		}
-
-		private String getMethodPathAnnotation() {
-			Path path = AnnotationUtils.findAnnotation(method, Path.class);
-			return path != null ? path.value() : "";
-		}
+		register(SimpleResource.class);
+		// when spring-boot 1.3.0 is released
+		// this can be done also via 'BaseJerseyConfig'
+		register(new BestMatchingPatternFilter("/api"));
 	}
 }
