--- samples/zmon-actuator-jetty-jersey/src/main/java/org/zalando/zmon/boot/jetty/controller/SimpleController.java
@@ -1,18 +1,0 @@
-package org.zalando.zmon.boot.jetty.controller;
-
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-
-import org.springframework.stereotype.Component;
-
-@Component
-@Path("/simple")
-public class SimpleController {
-
-	@GET
-	@Path("/{id}/complex")
-	public String invoke(@PathParam("id") String id){
-		return id;
-	}
-}
