+++ samples/zmon-actuator-jetty-jersey/src/test/resources/application.yml
@@ -1,0 +1,2 @@
+server:
+  servlet-path: /spring
