+++ zmon-actuator-jaxrs-jersey/pom.xml
@@ -13,9 +13,15 @@
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-jersey</artifactId>
 		</dependency>
+		<!-- DEFAULT VERSION BY BOOT DOES NOT WORK -->
 		<dependency>
 			<groupId>org.glassfish.jersey.core</groupId>
 			<artifactId>jersey-server</artifactId>
+			<version>2.22.1</version>
+		</dependency>
+		<dependency>
+			<groupId>org.glassfish.jersey.containers</groupId>
+			<artifactId>jersey-container-servlet-core</artifactId>
 			<version>2.22.1</version>
 		</dependency>
 	</dependencies>
