+++ zmon-actuator/src/main/resources/META-INF/spring.factories
