+++ zmon-actuator/src/test/java/org/zalando/zmon/actuator/ExampleApplication.java
